refactor(errors): renamed some error files to match their class name
