+++ src/errors/error-at-index.ts
